refactor: 디렉터리 위치 이동
